refactor server/routes/helpers and server/utils/ into server/utils/ with more separate files
