+++ server/utils/SanitizeHex.ts
@@ -1,0 +1,11 @@
+export function sanitizeHex(str: string) {
+  if (typeof str !== 'string') return false;
+  if (str.length & 1) return false;
+  for (let i = 0; i < str.length; ++i) {
+    const ch = str.charAt(i);
+    if (!/[0-9A-Fa-f]/.test(ch)) {
+      return false;
+    }
+  }
+  return true;
+}
