+++ src/types.ts
@@ -1,0 +1,3 @@
+export type ActionResult =
+  | { ok: true; json: any }
+  | { ok: false; errorNumber: number };
